Moving IndexOfVectorized (and tests) to System.Slices as part of cleanup
