+++ src/System.Buffers.Experimental/System.Buffers.Experimental.csproj
@@ -13,6 +13,5 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="System.Buffers" Version="4.3.0" />
-    <PackageReference Include="System.Numerics.Vectors" Version="4.3.0" />
   </ItemGroup>
 </Project>
